+++ src/problem/Assignment.kt
@@ -1,6 +1,8 @@
 package problem
 
 import data.AssignmentCfg
+import data.QuadraticAssignment
+import data.combinations
 import data.newName
 import discrete.Domain
 import discrete.Goal
@@ -17,6 +19,8 @@ fun newAssignment(variant: String, n: Int): Problem? {
 	return when (variant) {
 		"basic" -> assignment(name)
 		"bottleneck" -> bottleneckAssignment(name)
+		"quadratic" -> quadraticAssignment(name)
+		"quadratic_bottleneck" -> quadraticBottleneckAssignment(name)
 		else -> null
 	}
 }
@@ -72,6 +76,58 @@ fun bottleneckAssignment(name: String): Problem? {
 	p.objectiveFn = fun(solution: Solution): Score {
 		if(solution.map.isEmpty()) return 0.0
 		return solution.map.entries.maxOf { (worker, task) -> cfg.cost[worker][task] }
+	}
+	return p
+}
+
+fun newQuadraticAssignmentProblem(name: String): Pair<Problem?, QuadraticAssignment?> {
+	val cfg = QuadraticAssignment.new(name) ?: return Pair(null, null)
+	val p = Problem(
+		name,
+		description = cfg.toString(),
+		type = ProblemType.SEQUENCE,
+		goal = Goal.MINIMIZE,
+		variables = Variables.index(cfg.count),
+	)
+	p.addVariableDomains(Domain.index(cfg.count))
+	p.addUniversalConstraint(Constraint::allUnique)
+	p.solutionStringFn = StringFn.sequence((1..cfg.count).toList())
+	return Pair(p, cfg)
+}
+
+fun quadraticAssignment(name: String): Problem? {
+	val (p, cfg) = newQuadraticAssignmentProblem(name)
+	if (p == null || cfg == null) return null
+
+	p.objectiveFn = fun(solution: Solution): Score {
+		var totalCost: Score = 0.0
+		for(pair in p.variables.combinations(2)) {
+			val (facility1, facility2) = pair
+			val location1 = solution.map[facility1] ?: continue
+			val location2 = solution.map[facility2] ?: continue
+			totalCost += cfg.flow[facility1][facility2] * cfg.distance[location1][location2]
+			totalCost += cfg.flow[facility2][facility1] * cfg.distance[location2][location1]
+		}
+		return totalCost
+	}
+	return p
+}
+
+fun quadraticBottleneckAssignment(name: String): Problem? {
+	val (p, cfg) = newQuadraticAssignmentProblem(name)
+	if (p == null || cfg == null) return null
+
+	p.objectiveFn = fun(solution: Solution): Score {
+		var maxCost: Score = 0.0
+		for(pair in p.variables.combinations(2)) {
+			val (facility1, facility2) = pair
+			val location1 = solution.map[facility1] ?: continue
+			val location2 = solution.map[facility2] ?: continue
+			val cost1 = cfg.flow[facility1][facility2] * cfg.distance[location1][location2]
+			val cost2 = cfg.flow[facility2][facility1] * cfg.distance[location2][location1]
+			maxCost = listOf(maxCost, cost1, cost2).max()
+		}
+		return maxCost
 	}
 	return p
 }